szimulációs lépés

+++ week09_mikroszim/week09_mikroszim/Entities/DeathProbability.cs
@@ -9,7 +9,7 @@ namespace week09_mikroszim.Entities
     public class DeathProbability
     {
         public Gender Gender { get; set; }
-        public int Age { get; set; }
+        public byte Age { get; set; }
         public double P { get; set; }
 
     }
